Refactor CurrentSiteContext to utilize getCurrentSiteInfo utility and update PPAPIService to fetch website ID accordingly
