+++ src/client/power-pages/actions-hub/CurrentSiteContext.ts
@@ -4,13 +4,8 @@
  */
 
 import * as vscode from 'vscode';
-import * as path from 'path';
-import { findWebsiteYmlFolder, getWebsiteRecordId } from '../../../common/utilities/WorkspaceInfoFinderUtil';
-
-interface ICurrentSiteContext {
-    currentSiteId: string | null;
-    currentSiteFolderPath: string | null;
-}
+import { getCurrentSiteInfo } from '../../../common/utilities/Utils';
+import { ICurrentSiteContext } from '../../../common/constants';
 
 class CurrentSiteContext implements ICurrentSiteContext {
     private _currentSiteId: string | null = null;
@@ -20,13 +15,13 @@ class CurrentSiteContext implements ICurrentSiteContext {
     public readonly onChanged = this._onChanged.event;
 
     constructor() {
-        const { currentSiteId, currentSiteFolderPath } = this.getCurrentSiteInfo();
+        const { currentSiteId, currentSiteFolderPath } = getCurrentSiteInfo();
         this._currentSiteId = currentSiteId;
         this._currentSiteFolderPath = currentSiteFolderPath;
 
         this._disposables.push(
             vscode.window.onDidChangeActiveTextEditor(() => {
-                const newSiteInfo = this.getCurrentSiteInfo();
+                const newSiteInfo = getCurrentSiteInfo();
                 if (newSiteInfo.currentSiteId !== this._currentSiteId) {
                     this._currentSiteId = newSiteInfo.currentSiteId;
                     this._currentSiteFolderPath = newSiteInfo.currentSiteFolderPath;
@@ -47,30 +42,6 @@ class CurrentSiteContext implements ICurrentSiteContext {
     public dispose() {
         this._onChanged.dispose();
         this._disposables.forEach(d => d.dispose());
-    }
-
-    private getCurrentSiteInfo(): ICurrentSiteContext {
-        const activeEditor = vscode.window.activeTextEditor;
-        if (activeEditor) {
-            const filePath = activeEditor.document.uri.fsPath;
-            const fileDirectory = path.dirname(filePath);
-            const websiteYmlFolder = findWebsiteYmlFolder(fileDirectory);
-
-            if (websiteYmlFolder) {
-                const siteId = getWebsiteRecordId(websiteYmlFolder);
-                if (siteId) {
-                    return { currentSiteId: siteId, currentSiteFolderPath: websiteYmlFolder };
-                }
-            }
-        }
-
-        // Fallback: check first workspace folder
-        const workspaceFolders = vscode.workspace.workspaceFolders;
-        if (workspaceFolders && workspaceFolders.length > 0) {
-            return { currentSiteId: getWebsiteRecordId(workspaceFolders[0].uri.fsPath), currentSiteFolderPath: workspaceFolders[0].uri.fsPath };
-        }
-
-        return { currentSiteId: null, currentSiteFolderPath: null };
     }
 }
